Add files via upload
# Introduction
This repository contains the code and analysis for Task 2 of my Data Science internship at SkillCraft Technology. For this task, I performed data cleaning and exploratory data analysis (EDA) on a Titanic dataset.

#c Dataset
The Titanic dataset is a well-known dataset in the field of data science. This dataset is widely used for educational purposes and provides an opportunity to explore various data analysis techniques.

# Task Description
- The main objective of this task was to perform data cleaning and exploratory data analysis.
- Explore the relationships between variables and identify patterns and trends in the data.

#Observation
The data cleaning and exploratory data analysis performed on the Titanic dataset provided valuable insights into the factors influencing survival rates during the Titanic disaster. This analysis contributes to a better understanding of historical events and showcases the application of data science techniques in deriving meaningful insights from data.


## 🤝 Let's Connect !!
Dive into the repository, share your thoughts, or reach out for discussions on data science, internship experiences, or any related topics. I'm excited to learn and grow together!

## 📧  Contact Information
- For questions, feedback, or collaboration opportunities, feel free to reach out. Looking forward to connecting!
- Email: aabshaarshaikh@gmail.com
- LinkedIn: www.linkedin.com/in/aabshaar-shaikh-046572261
### Thank you for your interest in this project. Happy coding! 🚀